cracker: change the type of variable pin_count from float to int

+++ src/cracker.h
@@ -48,6 +48,6 @@
 
 void crack();
 void advance_pin_count();
-void display_status(float pin_count, time_t start_time);
+void display_status(int pin_count, time_t start_time);
 
 #endif
